requirements.txt
torch>=2.0.0
transformers>=4.40.0
peft>=0.5.0
accelerate>=0.25.0
wandb
pillow
tqdm
numpy
diffusers
openpyxl
